// cosmowcl/src/app/shared/models/api/cosmopratiche/pratica.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * Pratiche
 * API per la gestione pratiche 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Processo } from './processo.d';
import { RiferimentoFruitore } from './riferimentoFruitore.d';
import { Attivita } from './attivita.d';
import { Variabile } from './variabile.d';
import { TipoPratica } from './tipoPratica.d';
import { CondivisionePratica } from './condivisionePratica.d';
import { StatoPratica } from './statoPratica.d';


export interface Pratica { 
    id?: number;
    codiceIpaEnte?: string;
    tipo?: TipoPratica;
    oggetto?: string;
    stato?: StatoPratica;
    riassunto?: string;
    idPraticaExt?: string;
    linkPratica?: string;
    linkPraticaEsterna?: string;
    uuidNodo?: string;
    utenteCreazionePratica?: string;
    dataCreazionePratica?: string;
    dataFinePratica?: string;
    dataCambioStato?: string;
    preferita?: boolean;
    dataAggiornamentoPratica?: string;
    attivita?: Array<Attivita>;
    processo?: Processo;
    fruitore?: RiferimentoFruitore;
    metadati?: string;
    dataCancellazione?: string;
    condivisioni?: Array<CondivisionePratica>;
    visibilita?: Array<string>;
    associata?: boolean;
    esterna?: boolean;
    tags?: Array<number>;
    variabiliProcesso?: Array<Variabile>;
}